requirements.txt
Django==5.0.*
psycopg2-binary
gunicorn
Pillow
httpx
python-dotenv
django-environ
pytils